GlobalErrorResponseDTO + uitgebreide GlobalExceptionHandler (DuplicateEmailException-bug gefixt, validatiefouten netjes afgehandeld), wachtwoord-sterkte-validatie toegevoegd

+++ Backend/src/main/java/com/StefanKiers/ToDoApp/BulletJournal/ToDoApp/repository/ProjectMemberRepository.java
@@ -15,5 +15,5 @@ import java.util.Optional;
 
 public interface ProjectMemberRepository extends JpaRepository<ProjectMember, Integer> {
   List<ProjectMember> findByProjectId(Integer projectId);
-  Optional<ProjectMember> findByProjectIdAndUserId(Integer projectId, Long userId);
+  Optional<ProjectMember> findByProjectIdAndUserId(Long projectId, Long userId);
 }
